+++ x/bank/keeper/genesis.go
@@ -30,10 +30,6 @@ func (k BaseKeeper) InitGenesis(ctx context.Context, genState *types.GenesisStat
 
 	for _, balance := range genState.Balances {
 		addr := balance.GetAddress()
-		bz, err := k.ak.AddressCodec().StringToBytes(addr)
-		if err != nil {
-			return err
-		}
 
 		for _, coin := range balance.Coins {
 			err := k.Balances.Set(ctx, collections.Join(sdk.AccAddress(bz), coin.Denom), coin.Amount)