+++ x/bank/types/balance.go
@@ -3,7 +3,6 @@ package types
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 	"sort"
 
 	"cosmossdk.io/core/address"
@@ -16,8 +15,8 @@ import (
 var _ exported.GenesisBalance = (*Balance)(nil)
 
 // GetAddress returns the account address of the Balance object.
-func (b Balance) GetAddress() string {
-	return b.Address
+func (b Balance) GetAddress() sdk.AccAddress {
+	return sdk.MustAccAddressFromBech32(b.Address)
 }
 
 // GetCoins returns the account coins of the Balance object.